Created GUI with TableView elements for the SqlPlayer Class data
Updated the DatabaseMain class.

Created Table.fxml with the general GUI elements; can be opened with Scene Builder for easy GUI editing.

Created the TableController class which handles most of the Table processing.

Created the PlayerTable Class which has the constructor, setters and getters for retrieving and updating data from the SqlPlayer class.

Created the DBConnector class to store the SQL connection string.

NOTE: If the program doesn't run correctly make sure that you have the run configuration vm argument configured and that the appropriate javaFX build paths are installed. If not I refer to this video: https://www.youtube.com/watch?v=bC4XB6JAaoU&t=323s

Also, if any SQLExceptions occur, mess around with the execution statements in the TableController class.